+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Overwrite the `scale` parameter when running baysor and providing both `config` and `scale` parameter (#294)
 - Avoid loosing prior cell `0` in prior assignment when `unassigned_value != 0`
+- Fixed stardist dependencies, as it still doesn't support `numpy>=2.0.0`
 
 ## [2.1.2] - 2025-08-16